requirements.txt
timm==0.4.9
transformers
ruamel_yaml
opencv-python
scikit-image
matplotlib
pycocotools
pycocoevalcap
datasets
